--- Exception/UnableToInitializeSchemaServiceException.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Youshido\GraphQLBundle\Exception;
-
-use Exception;
-
-class UnableToInitializeSchemaServiceException extends Exception
-{
-}
