+++ src/Exception/UnableToInitializeSchemaServiceException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Youshido\GraphQLBundle\src\Exception;
+
+use Exception;
+
+class UnableToInitializeSchemaServiceException extends Exception
+{
+}
